batch_chem demo driver: replace getopt with petsc command line processing.
Using getopt prevents use of the petsc or mpi command line options for
debugging (malloc/memory issues). Even though this is a serial program
that uses very little of petsc or mpi, this is an important option to
keep available. Also avoids licensing issues because xgetopt claims to
be in the public domain.