undefined behavior in pflotran interface
PFloTran::RReact() never sets ierror to 0 -- if all goes well, it does not touch ierror.  It only ever sets it to 1 if convergence fails.

As a result, line 553 is undefined behavior.  Under the rare but possible case that ierror is "not initialized" to 1, this takes the wrong branch of the code, but either way undefined behavior is a bad thing.

This bug was found via valgrind...